implement typechecking for terms
